Level transition fix
using @export var level_scene: PackedScene seems to be having issues with uniqueness. Godot is adding on extra numbers to the end of the file name and then not finding the scene after it transitions successfully the first time. Sticking with hard coded scene paths in the load_scene_from_file() function for now.